requirements.txt
pyannote.audio
pydub
openai
python-dotenv
torch
torchvision
torchaudio
langchain-core
langchain-community
langchain-openai